% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{reorder_within}
\alias{reorder_within}
\title{Helper: Reorder Terms Within Facets}
\usage{
reorder_within(x, by, within, sep = "___")
}
\description{
Helper: Reorder Terms Within Facets
}
\keyword{internal}
