% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scale_y_reordered}
\alias{scale_y_reordered}
\title{Helper: Scale for Reordered Terms}
\usage{
scale_y_reordered(..., sep = "___")
}
\description{
Helper: Scale for Reordered Terms
}
\keyword{internal}
