% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{SportMiner-package}
\alias{SportMiner}
\alias{SportMiner-package}
\title{SportMiner: Text Mining and Topic Modeling for Sport Science Literature}
\description{
A comprehensive toolkit for mining, analyzing, and visualizing scientific literature in sport science domains. Provides functions for retrieving abstracts from 'Scopus', preprocessing text data, performing advanced topic modeling using Latent Dirichlet Allocation ('LDA'), Structural Topic Models ('STM'), and Correlated Topic Models ('CTM'), and creating publication-ready visualizations including keyword co-occurrence networks and topic trends. For methodological details see Blei et al. (2003) \doi{10.1162/jmlr.2003.3.4-5.993} for 'LDA', Roberts et al. (2014) \doi{10.1111/ajps.12103} for 'STM', and Blei and Lafferty (2007) \doi{10.1214/07-AOAS114} for 'CTM'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/praveenmaths89/SportMiner}
  \item Report bugs at \url{https://github.com/praveenmaths89/SportMiner/issues}
}

}
\author{
\strong{Maintainer}: Praveen D Chougale \email{praveenmaths89@gmail.com}

Authors:
\itemize{
  \item Usha Ananthakumar \email{usha@som.iitb.ac.in}
}

}
\keyword{internal}
